stream_reader/requirements.txt
kafka-python==2.0.1
psycopg2==2.9.3